#1327 Step 1: split poll_descriptor.rs (Phase 1.5 follow-up to #946)
Closes #1327.

Convert flat userspace-dp/src/afxdp/poll_descriptor.rs (3292 LOC,
single 2490-LOC `poll_binding_process_descriptor`) into a directory
module and extract two self-contained blocks. This is #946 Phase 1.5
— the helper-extraction follow-up that #946's closure note left
open. It is pure code motion: every continue site, every push, and
every side effect inside the lifted blocks is preserved verbatim.

Layout after this PR:
  userspace-dp/src/afxdp/poll_descriptor/
    mod.rs              2806 LOC — pub(super) fn driver + small
                        pre-existing helpers + slow-path body.
    flow_cache_hit.rs    379 LOC — stage_flow_cache_hit
                        (#[inline(always)]) + FlowCacheOutcome
                        { Consumed, FallThrough } enum.
    rx_telemetry.rs      220 LOC — record_rx_descriptor_telemetry
                        verbatim from #1128.

Hot-path discipline:
  - #[inline(always)] on stage_flow_cache_hit. Verified by
    `nm /dev/shm/cargo/release/xpf-userspace-dp | grep
    stage_flow_cache_hit` (no symbol) and `objdump -d | grep`
    (no call edge): LLVM fully inlined the helper.
  - Zero new per-packet allocations. Every cached_* binding is
    a borrow into the existing flow-cache Cache entry. The only
    heap interactions in the helper are pre-existing
    (owned_packet_frame.take() + PendingForwardRequest build).
  - Helper owns ALL scratch_recycle.push and scratch_forwards.push
    sites for the flow-cache fast path. On FlowCacheOutcome::Consumed
    the caller MUST `continue` without touching desc.addr again —
    contract documented in flow_cache_hit.rs preamble.

Borrow-safety:
  - Helper takes 8 disjoint binding sub-struct refs/scalars
    (&mut binding.flow, &mut binding.tx_pipeline,
    &mut binding.tx_counters, &mut binding.scratch,
    &mut binding.mirror_sample_counter, &binding.live,
    binding.slot, binding_index), NOT &mut BindingWorker.
    Disjoint from `binding.xsk.rx` which is held by the outer
    `received` binding.
  - Helper takes &mut owned_packet_frame and re-derives
    packet_frame internally. Caller's L477 packet_frame
    binding's NLL lifetime ends at L546 (last pre-flow-cache
    use); caller re-binds packet_frame after the helper for the
    L880+ slow-path code.

Plan-review history:
  - Round 1 (v1): Codex + AGY both PLAN-NEEDS-MAJOR.
  - Round 2 (v2): Codex + AGY both PLAN-NEEDS-MINOR.
  - Round 3 (v3): Codex PLAN-NEEDS-MAJOR (caught packet_frame /
    owned_packet_frame borrow hazard); AGY PLAN-READY.
  - Round 4 (v4): AGY PLAN-READY; Codex provisional PLAN-READY
    (sandbox infra blocked direct filesystem verification).
  - Task IDs recorded in docs/pr/1327-poll-descriptor-stages/reviewer-ids.md.

Stages 12+ verdict:
  Five candidates evaluated (resolve_session, nat_preroute,
  policy_eval, nat_postroute, fib_neighbor_resolve,
  session_install). All blocked by mutable-locals coupling that
  would need a separate #946-Phase-2-class DescriptorState plan.
  Step 1 is the structural ceiling for #1327; closing the issue
  after this merges is correct.

Validation:
  - cargo build --release: clean (114 pre-existing warnings).
  - cargo test --release: 1417/1417 main suite pass. The
    snat_contract_doc_guard test fails identically on master and
    branch (docs/userspace-dataplane-gaps.md missing the
    "fail-closed" keyword); pre-existing, unrelated to #1327.
    Test updated only to accept the directory layout.
  - 5/10 afxdp:: flake: same 2/10 wg::engine concurrent-load
    failure on master and branch; pre-existing, unrelated. Solo,
    the failing test passes 10/10 on both master and branch.
  - cargo asm gate: no `call` to stage_flow_cache_hit emitted
    (nm + objdump confirm).
  - Go suite: 30/30 packages pass.
